requirements.txt
Django==1.11
django-bootstrap-themes==3.3.6
django-datetime-widget==0.9.3

# for test and dev
PyHamcrest==1.9.0
ipdb==0.10.3
ipython==6.0.0
